refact : variant index
