+++ app/GraphQL/Inventory/Builders/Variants/VariantChannelBuilder.php
@@ -1,0 +1,61 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\GraphQL\Inventory\Builders\Variants;
+
+use GraphQL\Type\Definition\ResolveInfo;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Facades\DB;
+use Kanvas\Inventory\Channels\Models\Channels;
+use Kanvas\Inventory\Variants\Models\Variants as ModelsVariants;
+use Kanvas\Inventory\Variants\Models\VariantsChannels;
+use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
+
+class VariantChannelBuilder
+{
+    public function allVariantsPublishedInChannel(
+        mixed $root,
+        array $args,
+        GraphQLContext $context,
+        ResolveInfo $resolveInfo
+    ): Builder {
+        $channelUuid = $args['id'];
+
+        $channel = Channels::getByUuid($channelUuid);
+        $variants = new ModelsVariants();
+        $variantsChannel = new VariantsChannels();
+
+        //set index
+        ModelsVariants::setSearchIndex((int) $channel->companies_id);
+
+        /**
+         * @var Builder
+         */
+        return ModelsVariants::select(
+            $variants->getTable() . '.*',
+            DB::raw("'{$channel->name}' as channel_name"), //add channel name
+            $variantsChannel->getTable() . '.price',
+            $variantsChannel->getTable() . '.discounted_price',
+            $variantsChannel->getTable() . '.is_published',
+        )
+        ->join($variantsChannel->getTable(), $variantsChannel->getTable() . '.products_variants_id', '=', $variants->getTable() . '.id')
+        ->where($variantsChannel->getTable() . '.channels_id', $channel->getId())
+        ->where($variantsChannel->getTable() . '.is_deleted', 0)
+        ->where($variantsChannel->getTable() . '.is_published', 1);
+    }
+
+    /**
+     * Format channel data from builder
+     */
+    public function getChannel(mixed $root, array $req): array
+    {
+        return [
+            'name' => $root->channel_name,
+            'price' => $root->price,
+            'warehouses_id' => 0, //remove -_-
+            'discounted_price' => $root->discounted_price,
+            'is_published' => $root->is_published,
+        ];
+    }
+}
