+++ app/GraphQL/Inventory/Builders/Variants/VariantWarehouseBuilder.php
@@ -6,48 +6,14 @@ namespace App\GraphQL\Inventory\Builders\Variants;
 
 use GraphQL\Type\Definition\ResolveInfo;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Support\Facades\DB;
-use Kanvas\Inventory\Channels\Models\Channels;
 use Kanvas\Inventory\Status\Models\Status;
 use Kanvas\Inventory\Variants\Models\Variants as ModelsVariants;
-use Kanvas\Inventory\Variants\Models\VariantsChannels;
 use Kanvas\Inventory\Variants\Models\VariantsWarehouses;
 use Kanvas\Inventory\Warehouses\Models\Warehouses;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
-class Variants
+class VariantWarehouseBuilder
 {
-    public function allVariantsPublishedInChannel(
-        mixed $root,
-        array $args,
-        GraphQLContext $context,
-        ResolveInfo $resolveInfo
-    ): Builder {
-        $channelUuid = $args['id'];
-
-        $channel = Channels::getByUuid($channelUuid);
-        $variants = new ModelsVariants();
-        $variantsChannel = new VariantsChannels();
-
-        //set index
-        ModelsVariants::setSearchIndex((int) $channel->companies_id);
-
-        /**
-         * @var Builder
-         */
-        return ModelsVariants::select(
-            $variants->getTable() . '.*',
-            DB::raw("'{$channel->name}' as channel_name"), //add channel name
-            $variantsChannel->getTable() . '.price',
-            $variantsChannel->getTable() . '.discounted_price',
-            $variantsChannel->getTable() . '.is_published',
-        )
-        ->join($variantsChannel->getTable(), $variantsChannel->getTable() . '.products_variants_id', '=', $variants->getTable() . '.id')
-        ->where($variantsChannel->getTable() . '.channels_id', $channel->getId())
-        ->where($variantsChannel->getTable() . '.is_deleted', 0)
-        ->where($variantsChannel->getTable() . '.is_published', 1);
-    }
-
     public function allVariantsInWarehouse(
         mixed $root,
         array $args,
@@ -73,20 +39,6 @@ class Variants
             ->where($variantWarehouse->getTable() . '.warehouses_id', $warehouse->getId())
             ->where($variantWarehouse->getTable() . '.is_deleted', 0)
             ->where($variantWarehouse->getTable() . '.is_published', 1);
-    }
-
-    /**
-     * Format channel data from builder
-     */
-    public function getChannel(mixed $root, array $req): array
-    {
-        return [
-            'name' => $root->channel_name,
-            'price' => $root->price,
-            'warehouses_id' => 0, //remove -_-
-            'discounted_price' => $root->discounted_price,
-            'is_published' => $root->is_published,
-        ];
     }
 
     public function getVariantsByStatus(
